+++ backend/canisters/user/impl/src/model/direct_chats.rs
@@ -59,16 +59,8 @@ impl DirectChats {
             .collect()
     }
 
-    pub fn pinned(&self) -> &HashMap<ChatId, TimestampMillis> {
-        &self.pinned.value
-    }
-
     pub fn pinned_chats(&self) -> HashMap<Chat, TimestampMillis> {
         self.pinned.value.iter().map(|(k, v)| (Chat::Direct(*k), *v)).collect()
-    }
-
-    pub fn pinned_if_updated(&self, since: TimestampMillis) -> Option<HashMap<ChatId, TimestampMillis>> {
-        self.pinned.if_set_after(since).map(|ids| ids.to_owned())
     }
 
     pub fn pinned_chats_if_updated(&self, since: TimestampMillis) -> Option<HashMap<Chat, TimestampMillis>> {