+++ backend/canisters/user/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Removed pinned from direct and group chat initial_state and updates ([#8631](https://github.com/open-chat-labs/open-chat/pull/8631))
+
 ## [[2.0.1908-user](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1908-user)] - 2025-09-24
 
 ### Changed